+++ app/graphql/mutations/companies/create_company.rb
@@ -1,14 +1,14 @@
 # typed: true
 class Mutations::Companies::CreateCompany < Mutations::BaseMutation
-  description "Create a new game company. **Only available when using a first-party OAuth Application.**"
+  description "Create a new game company. **Only available to moderators and admins using a first-party OAuth Application.**"
 
   argument :name, String, required: true, description: 'The name of the company.'
-  argument :wikidata_id, ID, required: false, description: 'The ID of the company item in Wikidata.'
+  argument :wikidata_id, ID, required: true, description: 'The ID of the company item in Wikidata.'
 
   field :company, Types::CompanyType, null: true, description: "The company that was created."
 
-  sig { params(name: String, wikidata_id: T.nilable(T.any(String, Integer))).returns(T::Hash[Symbol, Company]) }
-  def resolve(name:, wikidata_id: nil)
+  sig { params(name: String, wikidata_id: T.any(String, Integer)).returns(T::Hash[Symbol, Company]) }
+  def resolve(name:, wikidata_id:)
     company = Company.new(name: name, wikidata_id: wikidata_id)
 
     raise GraphQL::ExecutionError, company.errors.full_messages.join(", ") unless company.save