+++ app/graphql/mutations/companies/update_company.rb
@@ -1,6 +1,6 @@
 # typed: true
 class Mutations::Companies::UpdateCompany < Mutations::BaseMutation
-  description "Update an existing game company. **Only available when using a first-party OAuth Application.**"
+  description "Update an existing game company. **Only available to moderators and admins using a first-party OAuth Application.**"
 
   argument :company_id, ID, required: true, description: 'The ID of the company record.'
   argument :name, String, required: false, description: 'The name of the company.'