+++ app/graphql/mutations/engines/create_engine.rb
@@ -1,14 +1,14 @@
 # typed: true
 class Mutations::Engines::CreateEngine < Mutations::BaseMutation
-  description "Create a new game engine. **Only available when using a first-party OAuth Application.**"
+  description "Create a new game engine. **Only available to moderators and admins using a first-party OAuth Application.**"
 
   argument :name, String, required: true, description: 'The name of the engine.'
-  argument :wikidata_id, ID, required: false, description: 'The ID of the engine item in Wikidata.'
+  argument :wikidata_id, ID, required: true, description: 'The ID of the engine item in Wikidata.'
 
   field :engine, Types::EngineType, null: true, description: "The engine that was created."
 
-  sig { params(name: String, wikidata_id: T.nilable(T.any(String, Integer))).returns(T::Hash[Symbol, Engine]) }
-  def resolve(name:, wikidata_id: nil)
+  sig { params(name: String, wikidata_id: T.any(String, Integer)).returns(T::Hash[Symbol, Engine]) }
+  def resolve(name:, wikidata_id:)
     engine = Engine.new(name: name, wikidata_id: wikidata_id)
 
     raise GraphQL::ExecutionError, engine.errors.full_messages.join(", ") unless engine.save