+++ app/graphql/mutations/genres/create_genre.rb
@@ -3,12 +3,12 @@ class Mutations::Genres::CreateGenre < Mutations::BaseMutation
   description "Create a new game genre. **Only available to moderators and admins using a first-party OAuth Application.**"
 
   argument :name, String, required: true, description: 'The name of the genre.'
-  argument :wikidata_id, ID, required: false, description: 'The ID of the genre item in Wikidata.'
+  argument :wikidata_id, ID, required: true, description: 'The ID of the genre item in Wikidata.'
 
   field :genre, Types::GenreType, null: true, description: "The genre that was created."
 
-  sig { params(name: String, wikidata_id: T.nilable(T.any(String, Integer))).returns(T::Hash[Symbol, Genre]) }
-  def resolve(name:, wikidata_id: nil)
+  sig { params(name: String, wikidata_id: T.any(String, Integer)).returns(T::Hash[Symbol, Genre]) }
+  def resolve(name:, wikidata_id:)
     genre = Genre.new(name: name, wikidata_id: wikidata_id)
 
     raise GraphQL::ExecutionError, genre.errors.full_messages.join(", ") unless genre.save