+++ app/graphql/mutations/series/create_series.rb
@@ -1,14 +1,14 @@
 # typed: true
 class Mutations::Series::CreateSeries < Mutations::BaseMutation
-  description "Create a new game series. **Only available when using a first-party OAuth Application.**"
+  description "Create a new game series. **Only available to moderators and admins using a first-party OAuth Application.**"
 
   argument :name, String, required: true, description: 'The name of the series.'
-  argument :wikidata_id, ID, required: false, description: 'The ID of the series item in Wikidata.'
+  argument :wikidata_id, ID, required: true, description: 'The ID of the series item in Wikidata.'
 
   field :series, Types::SeriesType, null: true, description: "The series that was created."
 
-  sig { params(name: String, wikidata_id: T.nilable(T.any(String, Integer))).returns(T::Hash[Symbol, Series]) }
-  def resolve(name:, wikidata_id: nil)
+  sig { params(name: String, wikidata_id: T.any(String, Integer)).returns(T::Hash[Symbol, Series]) }
+  def resolve(name:, wikidata_id:)
     series = Series.new(name: name, wikidata_id: wikidata_id)
 
     raise GraphQL::ExecutionError, series.errors.full_messages.join(", ") unless series.save