+++ app/models/genre.rb
@@ -17,7 +17,7 @@ class Genre < ApplicationRecord
 
   validates :wikidata_id,
     uniqueness: true,
-    allow_blank: true,
+    presence: true,
     numericality: {
       only_integer: true,
       greater_than: 0