+++ db/seeds/companies.rb
@@ -3,6 +3,7 @@ puts "Creating Companies..."
 
 20.times do
   Company.create!(
-    name: Faker::Game.unique.company
+    name: Faker::Game.unique.company,
+    wikidata_id: Faker::Number.unique.number(digits: 6)
   )
 end