+++ spec/factories/engines.rb
@@ -2,15 +2,12 @@
 FactoryBot.define do
   factory :engine do
     name { "Source Engine" }
+    wikidata_id { Faker::Number.number(digits: 6) }
 
     trait :game do
       after(:create) { |engine| create(:game_engine, engine: engine) }
     end
 
-    trait :wikidata_id do
-      wikidata_id { Faker::Number.number(digits: 6) }
-    end
-
-    factory :engine_with_everything, traits: [:game, :wikidata_id]
+    factory :engine_with_everything, traits: [:game]
   end
 end