+++ spec/factories/platforms.rb
@@ -2,15 +2,12 @@
 FactoryBot.define do
   factory :platform do
     name { "Xbox 360" }
+    wikidata_id { Faker::Number.number(digits: 6) }
 
     trait :game do
       after(:create) { |platform| create(:game_platform, platform: platform) }
     end
 
-    trait :wikidata_id do
-      wikidata_id { Faker::Number.number(digits: 6) }
-    end
-
-    factory :platform_with_everything, traits: [:game, :wikidata_id]
+    factory :platform_with_everything, traits: [:game]
   end
 end