+++ spec/policies/company_policy_spec.rb
@@ -8,14 +8,16 @@ RSpec.describe CompanyPolicy, type: :policy do
     let(:user) { create(:user) }
     let(:company) { create(:company) }
 
-    it "let's a user do everything except destroy the company" do
+    it "let's a normal user list, show, and search companies" do
       expect(company_policy).to permit_actions(
-        [:index, :show, :create, :new, :edit, :update, :search]
+        [:index, :show, :search]
       )
     end
 
-    it 'does not let a user destroy companies' do
-      expect(company_policy).to forbid_action(:destroy)
+    it "doesn't let a normal user create, update, or destroy companies" do
+      expect(company_policy).to forbid_actions(
+        [:create, :new, :edit, :update, :destroy]
+      )
     end
   end