+++ spec/requests/api/mutations/engines/create_engine_spec.rb
@@ -18,7 +18,7 @@ RSpec.describe "CreateEngine Mutation API", type: :request do
       GRAPHQL
     end
 
-    [:user, :moderator, :admin].each do |role|
+    [:moderator, :admin].each do |role|
       context "when the current user is a(n) #{role}" do
         let(:user) { create("confirmed_#{role}".to_sym) }
 
@@ -38,6 +38,17 @@ RSpec.describe "CreateEngine Mutation API", type: :request do
             }
           )
         end
+      end
+    end
+
+    context 'when the current user is a normal member' do
+      let(:user) { create(:confirmed_user) }
+
+      it "does not create a new engine" do
+        expect do
+          result = api_request(query_string, variables: { name: 'GoldSrc', wikidata_id: 123 }, token: access_token)
+          expect(result.to_h['errors'].first['message']).to eq("You aren't allowed to create an engine.")
+        end.not_to change(Engine, :count)
       end
     end
   end