+++ spec/requests/api/mutations/engines/update_engine_spec.rb
@@ -19,7 +19,7 @@ RSpec.describe "UpdateEngine Mutation API", type: :request do
       GRAPHQL
     end
 
-    [:user, :moderator, :admin].each do |role|
+    [:moderator, :admin].each do |role|
       context "when the current user is a(n) #{role}" do
         let(:user) { create("confirmed_#{role}".to_sym) }
 
@@ -42,6 +42,17 @@ RSpec.describe "UpdateEngine Mutation API", type: :request do
           expect(engine.reload.name).to eq('GoldSrc')
           expect(engine.reload.wikidata_id).to eq(123)
         end
+      end
+    end
+
+    context 'when the current user is a normal member' do
+      let(:user) { create(:confirmed_user) }
+
+      it "does not update the engine" do
+        expect do
+          result = api_request(query_string, variables: { engine_id: engine.id, name: 'GoldSrc', wikidata_id: 123 }, token: access_token)
+          expect(result.to_h['errors'].first['message']).to eq("You aren't allowed to update this engine.")
+        end.not_to change(engine.reload, :name).from('Source Engine')
       end
     end
   end